Refactor User model to App\Models
